Updated web processing steps

+++ processWebSubmissions.sh
@@ -3,20 +3,24 @@
 # If the node is busy give up running
 
 # Setup Madgraph-Pythia-Delphes
-source /nfs_scratch/dasu/2022-03-07/Madgraph-Pythia-Delphes-Studies/setup.sh
+source /nfs_scratch/dasu/2023-02-14/Madgraph-Pythia-Delphes-Studies/setup.sh
 
 # Loop over all input files received
-export inpdir=$HOME/public/physics535-data
+export inpdir=$HOME/public/Madgraph-Pythia-Delphes-Data
 datestr=`date +%Y%m%d-%H%M%S`
 export datadir=$inpdir/$datestr
 cd $inpdir
 for file in *.inp; do
-    # Create unique output data directory for this run
-    mkdir -p $datadir
-    mv $file $datadir
-    cd $datadir
-    stdout=`basename $file .inp`.out
-    stderr=`basename $file .inp`.err
-    nohup time python $mg5dir/bin/mg5_aMC $file < /dev/null > $stdout 2> $stderr&
-    cd $inpdir
+    if [ -f "$file" ]
+    then
+	echo "Processing $file..."
+	# Create unique output data directory for this run
+	mkdir -p $datadir
+	mv $file $datadir
+	cd $datadir
+	stdout=`basename $file .inp`.out
+	stderr=`basename $file .inp`.err
+	nohup time python $mg5dir/bin/mg5_aMC $file < /dev/null > $stdout 2> $stderr&
+	cd $inpdir
+    fi
 done
